organized package structure to follow GitHub API structure
