+++ src/utils/pagination.jl
@@ -1,16 +1,3 @@
-
-# Utility -------
-
-function github_obj_from_type(data::Dict)
-    t = get(data, "type", nothing)
-
-    if t == "User"
-        return User(data)
-    elseif t == "Organization"
-        return Organization(data)
-    end
-end
-
 # Pagination Helper Functions
 
 # Parses a Link header into the links it represents
